Cleanup some queries.
