+++ src-electron/rest/user-data.js
@@ -24,6 +24,7 @@
 const env = require('../util/env.js')
 const queryZcl = require('../db/query-zcl.js')
 const queryConfig = require('../db/query-config.js')
+const queryEndpointType = require('../db/query-endpoint-type.js')
 const querySession = require('../db/query-session.js')
 const queryPackage = require('../db/query-package.js')
 const validation = require('../validation/validation.js')
@@ -256,7 +257,7 @@ function httpGetInitialState(db) {
     let state = {}
 
     let statePopulators = []
-    let endpointTypes = queryConfig
+    let endpointTypes = queryEndpointType
       .selectAllEndpointTypes(db, sessionId)
       .then((rows) => {
         state.endpointTypes = rows
