+++ test/importexport.test.js
@@ -30,6 +30,7 @@ const querySession = require('../src-electron/db/query-session.js')
 const testUtil = require('./test-util.js')
 const testQuery = require('./test-query.js')
 const queryConfig = require('../src-electron/db/query-config.js')
+const queryEndpointType = require('../src-electron/db/query-endpoint-type.js')
 const util = require('../src-electron/util/util.js')
 
 let db
@@ -191,7 +192,7 @@ test(
     let sid = await querySession.createBlankSession(db)
     await importJs.importDataFromFile(db, testLightIsc, { sessionId: sid })
     expect(sid).not.toBeUndefined()
-    let endpointTypes = await queryConfig.selectAllEndpointTypes(db, sid)
+    let endpointTypes = await queryEndpointType.selectAllEndpointTypes(db, sid)
     expect(endpointTypes.length).toBe(3)
     expect(endpointTypes[0].name).toBe('Centralized')
     expect(endpointTypes[1].name).toBe('GreenPower')
@@ -214,7 +215,7 @@ test(
     sid = await querySession.createBlankSession(db)
     await importJs.importDataFromFile(db, testDoorLockIsc, { sessionId: sid })
     expect(sid).not.toBeUndefined()
-    let endpointTypes = await queryConfig.selectAllEndpointTypes(db, sid)
+    let endpointTypes = await queryEndpointType.selectAllEndpointTypes(db, sid)
     expect(endpointTypes.length).toBe(1)
     let endpoints = await queryConfig.selectAllEndpoints(db, sid)
     expect(endpoints.length).toBe(1)
@@ -241,7 +242,7 @@ test(
     sid = await querySession.createBlankSession(db)
     await importJs.importDataFromFile(db, haLightIsc, { sessionId: sid })
     expect(sid).not.toBeUndefined()
-    let endpointTypes = await queryConfig.selectAllEndpointTypes(db, sid)
+    let endpointTypes = await queryEndpointType.selectAllEndpointTypes(db, sid)
     expect(endpointTypes.length).toBe(2)
     let endpoints = await queryConfig.selectAllEndpoints(db, sid)
     expect(endpoints.length).toBe(2)
@@ -281,7 +282,7 @@ test(
       sessionId: sid,
     })
     expect(sid).not.toBeUndefined()
-    let endpointTypes = await queryConfig.selectAllEndpointTypes(db, sid)
+    let endpointTypes = await queryEndpointType.selectAllEndpointTypes(db, sid)
     expect(endpointTypes.length).toBe(1)
     let endpoints = await queryConfig.selectAllEndpoints(db, sid)
     expect(endpoints.length).toBe(1)
